+++ climaf/__init__.py
@@ -7,7 +7,7 @@ Climaf is documented at ReadTheDocs : http://climaf.readthedocs.org/
 
 __all__=[ "cache" , "classes", "clogging",   "dataloc" , "driver" , "netcdfbasics", "operators", "period",
           "site_settings"  , "standard_operators" , "projects" ]
-import posixpath, os
+import posixpath, os, sys
 
 version="0.6.1"
 
@@ -22,15 +22,13 @@ if not already_inited  :
     print "Climaf version = "+version
     #
     # Set default logging levels
-    clogging.clog(os.getenv("CLIMAF_LOG_LEVEL","error"))
+    clogging.clog(os.getenv("CLIMAF_LOG_LEVEL","warning"))
     clogging.clog_file(os.getenv("CLIMAF_LOGFILE_LEVEL","info"))
     #
     # Read and execute user config file
     conf_file=os.path.expanduser("~/.climaf")
     if os.path.isfile(conf_file) :
-        with open(conf_file) as fobj:
-            startup_file = fobj.read()
-            exec(startup_file)
+        execfile(conf_file,sys.modules['__main__'].__dict__)
     #    
     # Decide for cache location
     if site_settings.onCiclad :